# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "企业级后台管理系统接口文档，覆盖认证、用户、角色、菜单、部门、字典和审计等业务模块。",
        "title": "Go Server 后台管理 API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/system/depts": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "返回完整部门树结构，常用于组织架构管理页面。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "部门管理"
                ],
                "summary": "查询部门树",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/go-server_internal_response.Body"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/internal_modules_dept.DeptTreeNode"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/go-server_internal_response.Body"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "创建新的部门节点，可指定上级部门。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "部门管理"
                ],
                "summary": "创建部门",
                "parameters": [
                    {
                        "description": "部门创建参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_modules_dept.CreateDeptRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/go-server_internal_response.Body"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/internal_modules_dept.DeptResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/go-server_internal_response.Body"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/go-server_internal_response.Body"
                        }
                    }
                }
            }
        },
        "/system/depts/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "根据部门 ID 查询部门详细信息。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "部门管理"
                ],
                "summary": "获取部门详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "部门 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/go-server_internal_response.Body"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/internal_modules_dept.DeptResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/go-server_internal_response.Body"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/go-server_internal_response.Body"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/go-server_internal_response.Body"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "修改部门名称、负责人、联系方式、排序和状态等信息。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "部门管理"
                ],
                "summary": "更新部门",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "部门 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "部门更新参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_modules_dept.UpdateDeptRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/go-server_internal_response.Body"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/internal_modules_dept.DeptResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/go-server_internal_response.Body"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/go-server_internal_response.Body"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/go-server_internal_response.Body"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "根据部门 ID 删除部门节点。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "部门管理"
                ],
                "summary": "删除部门",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "部门 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/go-server_internal_response.Body"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "additionalProperties": {
                                                "type": "boolean"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/go-server_internal_response.Body"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/go-server_internal_response.Body"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/go-server_internal_response.Body"
                        }
                    }
                }
            }
        },
        "/system/depts/{id}/users": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "分页查询归属于指定部门的用户。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "部门管理"
                ],
                "summary": "查询部门下的用户列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "部门 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页条数",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/go-server_internal_response.Body"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/internal_modules_dept.DeptUserPageResult"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/go-server_internal_response.Body"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/go-server_internal_response.Body"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "go-server_internal_response.Body": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "data": {},
                "message": {
                    "type": "string"
                },
                "request_id": {
                    "type": "string"
                }
            }
        },
...
        "internal_modules_dept.CreateDeptRequest": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "email": {
                    "type": "string",
                    "example": "rd@example.com"
                },
                "leader": {
                    "type": "string",
                    "example": "李经理"
                },
                "name": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 1,
                    "example": "研发部"
                },
                "parent_id": {
                    "type": "integer",
                    "example": 0
                },
                "phone": {
                    "type": "string",
                    "example": "13800138001"
                },
                "sort": {
                    "type": "integer",
                    "example": 1
                },
                "status": {
                    "type": "integer",
                    "enum": [
                        0,
                        1
                    ],
                    "example": 1
                }
            }
        },
        "internal_modules_dept.DeptResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string",
                    "example": "2026-04-11T12:00:00Z"
                },
                "email": {
                    "type": "string",
                    "example": "rd@example.com"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "leader": {
                    "type": "string",
                    "example": "李经理"
                },
                "name": {
                    "type": "string",
                    "example": "研发部"
                },
                "parent_id": {
                    "type": "integer",
                    "example": 0
                },
                "phone": {
                    "type": "string",
                    "example": "13800138001"
                },
                "sort": {
                    "type": "integer",
                    "example": 1
                },
                "status": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "internal_modules_dept.DeptTreeNode": {
            "type": "object",
            "properties": {
                "children": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_modules_dept.DeptTreeNode"
                    }
                },
                "email": {
                    "type": "string",
                    "example": "rd@example.com"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "leader": {
                    "type": "string",
                    "example": "李经理"
                },
                "name": {
                    "type": "string",
                    "example": "研发部"
                },
                "parent_id": {
                    "type": "integer",
                    "example": 0
                },
                "phone": {
                    "type": "string",
                    "example": "13800138001"
                },
                "sort": {
                    "type": "integer",
                    "example": 1
                },
                "status": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "internal_modules_dept.DeptUserPageResult": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_modules_dept.DeptUserResponse"
                    }
                },
                "page": {
                    "type": "integer",
                    "example": 1
                },
                "page_size": {
                    "type": "integer",
                    "example": 10
                },
                "total": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "internal_modules_dept.DeptUserResponse": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string",
                    "example": "zhangsan@example.com"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "name": {
                    "type": "string",
                    "example": "张三"
                },
                "phone": {
                    "type": "string",
                    "example": "13800138000"
                },
                "status": {
                    "type": "integer",
                    "example": 1
                },
                "username": {
                    "type": "string",
                    "example": "zhangsan"
                }
            }
        },
        "internal_modules_dept.UpdateDeptRequest": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "email": {
                    "type": "string",
                    "example": "rd@example.com"
                },
                "leader": {
                    "type": "string",
                    "example": "李经理"
                },
                "name": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 1,
                    "example": "研发部"
                },
                "parent_id": {
                    "type": "integer",
                    "example": 0
                },
                "phone": {
                    "type": "string",
                    "example": "13800138001"
                },
                "sort": {
                    "type": "integer",
                    "example": 1
                },
                "status": {
                    "type": "integer",
                    "enum": [
                        0,
                        1
                    ],
                    "example": 1
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "请填写 Bearer Token，格式为：Bearer \u003ctoken\u003e",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}